Add fraction.h/cpp
